+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2025-07-24
+
+### Added
+- Gamechanger! Added web buttons to play/pause and stop playing
+- Added information about the current status on the display
+
+### Fixed
+- Better volume and PA handling
+
 ## [1.0.1] - 2025-07-23
 
 ### Fixed